NXS-6681: General fixes (#13)
* Upgrade to .NET 8 and enhance email functionalities

- Updated `SendTransactionBuySendDelayAsync` method to include a new parameter for transaction details and improved email body clarity.
- Upgraded target frameworks from `net6.0` to `net8.0` across multiple projects.
- Updated package references to the latest versions for better compatibility and security.
- Removed `HighRisk` properties from `BuyModels.cs` and `SellModels.cs`, indicating a shift in risk evaluation.
- Simplified client-side logic in `buy.js` and `sell.js` by removing references to `HighRisk`.
- Updated Dockerfile to use .NET 8.0 images for the latest runtime environment.

* Cleanup

* Update trust level checks in sell.js

Refactor conditions to use `trustLevel` instead of `type` for "NEW" and "TRUSTED". This change ensures the logic accurately reflects the intended use of `trustLevel` for displaying UI elements and modifying properties.

* Refactor trustLevel comparisons for type safety

Updated equality checks for the `trustLevel` variable from `==` to `===` to enhance type safety. This change affects the "NEW" and "TRUSTED" trust levels while maintaining the existing logic and functionality of the application.

* Remove Microsoft.NET.Sdk.Functions package reference

This commit removes the package reference for
`Microsoft.NET.Sdk.Functions` version `4.6.0`
from the `Nexus.Samples.MailService.csproj` file.
This change may reflect a transition to a different
SDK for Azure Functions to leverage newer versions
or functionalities provided by other included packages.